Reuse configs for GroovyExtension and GroovyGradleExtension
