+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/GroovyGradleExtension.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2020 DiffPlug
+ * Copyright 2016-2023 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,37 +15,15 @@
  */
 package com.diffplug.gradle.spotless;
 
-import java.util.Objects;
-
 import javax.inject.Inject;
 
-import com.diffplug.spotless.extra.groovy.GrEclipseFormatterStep;
-import com.diffplug.spotless.java.ImportOrderStep;
-
-public class GroovyGradleExtension extends FormatExtension {
+public class GroovyGradleExtension extends BaseGroovyExtension {
 	private static final String GRADLE_FILE_EXTENSION = "*.gradle";
 	static final String NAME = "groovyGradle";
 
 	@Inject
 	public GroovyGradleExtension(SpotlessExtension spotless) {
 		super(spotless);
-	}
-
-	public void importOrder(String... importOrder) {
-		addStep(ImportOrderStep.forGroovy().createFrom(importOrder));
-	}
-
-	public void importOrderFile(Object importOrderFile) {
-		Objects.requireNonNull(importOrderFile);
-		addStep(ImportOrderStep.forGroovy().createFrom(getProject().file(importOrderFile)));
-	}
-
-	public GroovyExtension.GrEclipseConfig greclipse() {
-		return new GroovyExtension.GrEclipseConfig(GrEclipseFormatterStep.defaultVersion(), this);
-	}
-
-	public GroovyExtension.GrEclipseConfig greclipse(String version) {
-		return new GroovyExtension.GrEclipseConfig(version, this);
 	}
 
 	@Override
